backend/pyproject.toml
[tool.poetry]
name = "events-api"
version = "0.1.0"
description = ""
authors = ["Pavlo Korshykov <pavel.korshikov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.7"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
psycopg = "^3.1.17"
dj-rest-auth = "^5.0.2"
django-anymail = "^10.2"
celery = "^5.3.6"
redis = "^5.0.1"
django-allauth = "^0.60.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
